Add the real fly specimen and its CC-BY-4.0 attribution
The user supplied fly.glb, so the hero panel now has a proper mesh instead of
procedural primitives. Copied to viz-fly/public/models/ and
duolingo-clone/public/models/, verified identical by md5
(34f48c57967d2f839f0b235a03103375, 729336 bytes).

CC-BY-4.0 REQUIRES attribution wherever the work is displayed, so the credit is now
rendered in the HUD both beside the specimen panel and in the attribution line, and
README records the full citation with a note that removing the model must remove the
credit with it. ATTRIBUTION.txt ships next to the asset.

Inspected the file so the animation work rests on facts rather than guesswork: 172
nodes, 57 meshes, 7 materials, 9 embedded textures, 6511 vertices, no skin and no
baked animation. It is static, but the parts are separate NAMED nodes, which is what
makes animation possible anyway: 4 wing nodes, 48 leg nodes (6 legs of several
segments), 2 eye nodes, a head and a body root. Bounding box is about 4765x5767x3527
units, so roughly 6000x too large and centred off-origin, hence fit at load time.

Steered the fly agent to load this instead of continuing the procedural material work
it was on, including the node map, the scale correction, and the requirement to keep
the procedural fly as a fallback so the page cannot break if the asset is missing.

+++ viz-fly/tools/verify_dev.py
@@ -96,7 +96,9 @@ FPS_JS = """() => new Promise((res) => {
 
 
 def rgb(png: bytes) -> np.ndarray:
-    return np.asarray(Image.open(BytesIO(png)).convert('RGB'), dtype=np.int16)
+    # int32, not int16: the luma weights below multiply by 299/587 and silently wrap in
+    # int16, which reports a fully lit frame as near black.
+    return np.asarray(Image.open(BytesIO(png)).convert('RGB'), dtype=np.int32)
 
 
 def luma(a: np.ndarray) -> np.ndarray:
@@ -118,7 +120,7 @@ def visibility(png: bytes, bg=BG, lit=LIT) -> dict:
     """Is the fly actually there, and how much of the panel does it own."""
     a = rgb(png)
     h, w, _ = a.shape
-    dist = np.abs(a - np.array(bg, dtype=np.int16)).sum(axis=2)
+    dist = np.abs(a - np.array(bg, dtype=np.int32)).sum(axis=2)
     l = luma(a)
     mask = l > lit
     ys, xs = np.nonzero(mask)
